1、customize exception 2、Data validity check 3、Data accuracy

+++ src/main/java/com/superhonor/shipment/service/impl/ShipmentItemServiceImpl.java
@@ -4,7 +4,9 @@ import com.superhonor.shipment.dao.ShipmentItemRepository;
 import com.superhonor.shipment.dto.ShipmentItemDTO;
 import com.superhonor.shipment.dto.ShipmentSplitDTO;
 import com.superhonor.shipment.entity.ShipmentItem;
+import com.superhonor.shipment.exception.UserFriendlyException;
 import com.superhonor.shipment.service.ShipmentItemService;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -36,7 +38,14 @@ public class ShipmentItemServiceImpl implements ShipmentItemService {
     @Transactional(rollbackFor = Exception.class)
     public void splitShipment(ShipmentSplitDTO splitDTO) {
         Optional <ShipmentItem> optional = itemRepository.findById(splitDTO.getShipmentItemId());
-        // TODO 校验原值和新值是否一致
+        List<ShipmentItemDTO> items = splitDTO.getItems();
+        final BigDecimal[] sum = {BigDecimal.ZERO};
+        items.forEach(e -> {
+            sum[0] = sum[0].add(e.getAmount());
+        });
+        if(sum[0].compareTo(optional.get().getAmount()) != 0) {
+            throw new UserFriendlyException("拆分后重量与拆分前不一致");
+        }
         itemRepository.deleteById(splitDTO.getShipmentItemId());
         this.saveAll(optional.get().getShipmentid(), splitDTO.getItems());
     }
@@ -46,6 +55,10 @@ public class ShipmentItemServiceImpl implements ShipmentItemService {
     public void mergeShipment(List <ShipmentItemDTO> items) {
         final BigDecimal[] b = {BigDecimal.ZERO};
         items.forEach(e -> {
+            Optional<ShipmentItem> optional = itemRepository.findById(e.getId());
+            if(optional.get().getAmount().compareTo(e.getAmount()) != 0) {
+                throw new UserFriendlyException(StringUtils.join("编号为：[", e.getId(), "]的信息数据不一致！"));
+            }
             itemRepository.deleteById(e.getId());
             b[0] = b[0].add(e.getAmount());
         });
